Better std structure for account
