+++ framework/contracts/account/src/contract.rs
@@ -10,8 +10,7 @@ use abstract_std::{
             AccountInfo, Config, WhitelistedModules, CONFIG, INFO, SUSPENSION_STATUS,
             WHITELISTED_MODULES,
         },
-        types::UpdateSubAccountAction,
-        ExecuteMsg, InstantiateMsg, QueryMsg,
+        ExecuteMsg, InstantiateMsg, QueryMsg, UpdateSubAccountAction,
     },
     objects::{gov_type::GovernanceDetails, ownership, AccountId},
     version_control::Account,
