+++ framework/packages/abstract-client/src/client.rs
@@ -207,7 +207,7 @@ impl<Chain: CwEnv> AbstractClient<Chain> {
                     )
                     .map_err(Into::into)?;
 
-                let manager_config: abstract_std::account::responses::ConfigResponse = chain
+                let manager_config: abstract_std::account::ConfigResponse = chain
                     .query(
                         &abstract_std::account::QueryMsg::Config {},
                         &app_config.manager_address,
