+++ framework/packages/abstract-std/src/native/ibc_client.rs
@@ -220,13 +220,12 @@ impl InstalledModuleIdentification {
                 let target_account_id = self.account_id.clone().ok_or(no_account_id_error)?;
                 let account_base = vc.account(&target_account_id, &deps.querier)?;
 
-                let module_info: account::responses::ModuleAddressesResponse =
-                    deps.querier.query_wasm_smart(
-                        account_base.into_addr(),
-                        &account::QueryMsg::ModuleAddresses {
-                            ids: vec![self.module_info.id()],
-                        },
-                    )?;
+                let module_info: account::ModuleAddressesResponse = deps.querier.query_wasm_smart(
+                    account_base.into_addr(),
+                    &account::QueryMsg::ModuleAddresses {
+                        ids: vec![self.module_info.id()],
+                    },
+                )?;
                 module_info
                     .modules
                     .first()
